block: add ComparePaths
Creates a sort order for Paths.

Remove sorting of patches from Diff - document that the user can do it
using ComparePaths if required.

Sort patches in TestDiff, ExampleDiff so we have deterministic output
to check against.
